Login & SignUp pages (#23)
Revamped Surafels work on the two pages. Adjusted CSS to make the pages
look like the design and added responsiveness. The mobile responsiveness
can use some work, but they are there as a placeholder for now.

NOTE: Navbar has to now be manually placed per page since the login &
sign-up pages do not have the navbar. This can be changed.